requirements.txt
Flask

python-keystoneclient==1.7.1
python-swiftclient==2.6.0
python-dateutil==2.5.3
cryptography==1.4
MySQL-python==1.2.5
